move multicluster capability to sql rest test case
